# Use `find` for `search_n` when n=1