--- a/benchmarks/src/search_n.cpp
+++ b/benchmarks/src/search_n.cpp
@@ -5,7 +5,6 @@
 #include <benchmark/benchmark.h>
 #include <cstddef>
 #include <cstdint>
-#include <limits>
 #include <vector>
 
 #include "skewed_allocator.hpp"