--- a/tests/std/tests/P0896R4_ranges_alg_search_n/test.cpp
+++ b/tests/std/tests/P0896R4_ranges_alg_search_n/test.cpp
@@ -100,6 +100,20 @@ struct instantiator {
 
         // trivial case: unit needle
         {
+            const auto result = ranges::search_n(range, 1, P{1, 42});
+            static_assert(same_as<decltype(result), const ranges::subrange<ranges::iterator_t<Fwd>>>);
+            assert(result.begin() == ranges::next(range.begin(), 1));
+            assert(result.end() == ranges::next(range.begin(), 2));
+        }
+        {
+            const auto result = ranges::search_n(ranges::begin(range), ranges::end(range), 1, P{1, 42});
+            static_assert(same_as<decltype(result), const ranges::subrange<ranges::iterator_t<Fwd>>>);
+            assert(result.begin() == ranges::next(range.begin(), 1));
+            assert(result.end() == ranges::next(range.begin(), 2));
+        }
+
+        // trivial case: unit needle with predicate
+        {
             const auto result = ranges::search_n(range, 1, 0, cmp, get_first);
             static_assert(same_as<decltype(result), const ranges::subrange<ranges::iterator_t<Fwd>>>);
             assert(result.begin() == ranges::next(range.begin(), 1));